review form input :content not working yet..

+++ db/seeds.rb
@@ -6,19 +6,45 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 puts "Cleaning database..."
-Restaurant.destroy_all
+Restaurant.destroy_all if Rails.env.development?
 
 puts "Creating restaurants..."
-lsdlp = { name: "La Strada de la Pizza", address: "10 Pizza St, Some Greek City A6 42N", category: "Pizzeria", phone_number: "33 33 55 55 55", reviews: 3 }
-pizza_fun = { name: "Pizza FAN", address: "42S Pizzastreet High St, Some Greek City JK9 Q42", category: "Pizzeria", phone_number: "44 44 66 66 66", reviews: 4 }
-resone = { name: "Pdsza FAN", address: "42S Pizzastreet High St, Some Greek City JK9 Q42", category: "Pizzeria", phone_number: "44 44 66 66 66", reviews: 4 }
-restwo = { name: "Pizza FuN", address: "42S Pizzastreet High St, Some Greek City JK9 Q42", category: "Pizzeria", phone_number: "44 44 66 66 66", reviews: 4 }
-resthree = { name: "Pizza F4N", address: "42S Pizzastreet High St, Some Greek City JK9 Q42", category: "Pizzeria", phone_number: "44 44 66 66 66", reviews: 4 }
-restaurants_test = [lsdlp, pizza_fun, resone, restwo, resthree]
-restaurants_test.each do |attributes|
-  restaurant = Restaurant.create!(attributes)
-  puts "Created #{restaurant.name}"
-end
+Restaurant.create!(
+  name: "C'est Bon C'est Belge",
+  address: "Rue du Bon-Secours, Brussel",
+  phone_number: "+321 23 45 67 89",
+  category: "belgian"
+)
+
+Restaurant.create!(
+  name: "Epicure",
+  address: "Faubourg Saint Honoré, Paris",
+  phone_number: "+331 23 45 67 89",
+  category: "french"
+)
+
+Restaurant.create!(
+  name: "Ristorante A Mano",
+  address: "Strausberger Platz, Berlin",
+  phone_number: "+491 23 45 67 89",
+  category: "italian"
+)
+
+Restaurant.create!(
+  name: "Teppanyaki Restaurant Sazanka",
+  address: "Ferdinand Bolstraat, Amsterdam",
+  phone_number: "+311 23 45 67 89",
+  category: "japanese"
+)
+
+Restaurant.create!(
+  name: "Sichuan Folk",
+  address: "Shoreditch, London",
+  phone_number: "+441 23 45 67 89",
+  category: "chinese"
+)
+  puts "Created"
+
 puts "Finished!"
 #require 'faker'
 #10.times
